Move Import form type to Core
