+++ src/Wallabag/CoreBundle/Controller/Import/BrowserController.php
@@ -7,8 +7,8 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Wallabag\CoreBundle\Controller\AbstractController;
+use Wallabag\CoreBundle\Form\Type\UploadImportType;
 use Wallabag\CoreBundle\Import\ImportInterface;
-use Wallabag\ImportBundle\Form\Type\UploadImportType;
 
 abstract class BrowserController extends AbstractController
 {
