+++ src/Wallabag/CoreBundle/Controller/Import/ReadabilityController.php
@@ -8,8 +8,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Wallabag\CoreBundle\Controller\AbstractController;
+use Wallabag\CoreBundle\Form\Type\UploadImportType;
 use Wallabag\CoreBundle\Import\ReadabilityImport;
-use Wallabag\ImportBundle\Form\Type\UploadImportType;
 use Wallabag\ImportBundle\Redis\Producer as RedisProducer;
 
 class ReadabilityController extends AbstractController
