+++ src/Wallabag/CoreBundle/Controller/Import/WallabagController.php
@@ -7,8 +7,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Wallabag\CoreBundle\Controller\AbstractController;
+use Wallabag\CoreBundle\Form\Type\UploadImportType;
 use Wallabag\CoreBundle\Import\ImportInterface;
-use Wallabag\ImportBundle\Form\Type\UploadImportType;
 
 /**
  * Define Wallabag import for v1 and v2, since there are very similar.
