+++ src/Wallabag/CoreBundle/Form/Type/UploadImportType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wallabag\ImportBundle\Form\Type;
+namespace Wallabag\CoreBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
